Adds proper testing of `update` method
As both the modbus and mqtt client are passed as argument, we need to
validate their behavior in the Update method. Used `mockery` to generate
the mock implementations, as they are quite involved. Added a number of
test cases to check how the update method responds for varying results
from the modbus server.