Change container setup for local testing
Change the docker setup such that we can now run the tests in them.
Basically this means that we add the `stretchr/testify` dependencies in
them such that these are available in the build image.

Also split up the docker image into a build and a final image, using
[multi-stage builds]. The linking between the images is assured via the
updated docker-compose file. I am not sure whether this is the best
approach, but it seems a pragmatic one.

Based the image now off of an alpine-variant, which makes the build
image also a bit lighter.

Adds user / group envs to the build image, such that this one does not
need to run as root. For the final image based on `scratch`, this seems
not possible though.

[mult-stage-builds]: https://docs.docker.com/develop/develop-images/multistage-build/