Run tests in travis CI, in docker image
Run the local test suite against a CI.

There are plenty of ways of running the (small) test setup against
travis; most of them point to running the tests inside the travis
environment themselves and afterwards build and image to push through.

This is not really what I'd like, as this means you're not really
testing the same setup. Here, we can build the image in docker-compose
and afterwards take this image to run the tests in.