Simplified golang setup
Fixes the travis CI build to no longer build everything in docker first
prior to running the tests. Instead, the tests are ran in the docker
image that travis already runs for us. This goes a bit against the idea
of having a single build process that runs throughout the entire flow,
but as we only build a static binary (inherent to golang), we can just
use goreleaser for creating the approapriate docker images afterwards.