Use goreleaser for generating the docker images
Follow-up of CI rebuild; use goreleaser to copy in the finished binary
afterwards in a scratch image.

Removing the earlier compose-based setup; building is done in the CI and
afterwards pushed by goreleaser upstream.

Dockerfile itself is pretty basic; copies in built binary and config.
Adds a `nobody` user for execution.