Fixes #37
When trying to do updates for coils in using the `CoilGroup.Update`, it
was assumed that the results from the modbus client were returned as an
array of `uint16`. However, they are in fact arrays of `byte`, so the
conversion scheme that was part of `CoilGroup.Update` was NOK and thus
was wrong in case of setups with more of 8 consecutive coils (including
my own).